北陌提出的计数器能直接调整当前计数、解决对话框设置电压后不立即生效

+++ Gigavolt/Block/Store/MemoryBankGVElectricElement.cs
@@ -21,7 +21,7 @@ namespace Game {
         public override uint GetOutputVoltage(int face) => m_voltage;
 
         public override bool Simulate() {
-            float voltage = m_voltage;
+            uint voltage = m_voltage;
             bool flag = false;
             bool flag2 = false;
             bool flag3 = false;
